Add player factory function and bug fixes
Created the player factory function and methods to generate random,
non duplicate coordinates for the computer player. Fixed a bug in the
ship placement functions where the loop was going further than expected.
This was due to the num variable being a string rather than a number, so
num + ship.length would be '45' rather than 9. Converted num to a number.